+++ src/main/java/edu/mum/cs/cs525/labs/exercises/project/ui/framework/model/BaseEntity.java
