+++ src/main/java/edu/mum/cs/cs525/labs/exercises/project/ui/framework/uitoolkit/TableRow.java
@@ -1,0 +1,6 @@
+package edu.mum.cs.cs525.labs.exercises.project.ui.framework.uitoolkit;
+
+public abstract class TableRow {
+
+    public abstract Object[] row();
+}
