+++ src/main/java/edu/mum/cs/cs525/labs/exercises/project/ui/framework/uitoolkit/composite/JTableView.java
@@ -1,0 +1,28 @@
+
+package edu.mum.cs.cs525.labs.exercises.project.ui.framework.uitoolkit.composite;
+
+
+import com.google.gson.annotations.SerializedName;
+import edu.mum.cs.cs525.labs.exercises.project.ui.framework.uitoolkit.visitor.Visitor;
+
+import java.util.List;
+
+@SuppressWarnings("unused")
+public class JTableView extends View{
+
+    @SerializedName("columns")
+    private List<String> columns;
+
+    public List<String> getColumns() {
+        return columns;
+    }
+
+    public void setColumns(List<String> columns) {
+        this.columns = columns;
+    }
+
+    @Override
+    public void accept(Visitor visitor){
+        visitor.visit(this);
+    }
+}
