+++ src/main/resources/bankui.json
@@ -1,0 +1,85 @@
+ {
+   "id" : "main_panel",
+   "type" : "JPanel",
+    "width" : 575,
+    "height" : 371,
+    "x":0,
+    "y": 0,
+    "children" : [
+      {
+        "id" : "add_personal_acc_btn",
+        "type" : "JButton",
+        "title" : "Add personal account",
+        "x":24,
+        "y": 20,
+        "width" : 192,
+        "height" : 33
+      },
+      {
+        "id" : "add_company_acc_btn",
+        "type" : "JButton",
+        "title" : "Add company account",
+        "x":240,
+        "y": 20,
+        "width" : 192,
+        "height" : 33
+      },
+      {
+        "id" : "deposit_btn",
+        "type" : "JButton",
+        "title" : "Deposit",
+        "x" : 468,
+        "y" : 104,
+        "width":96,
+        "height": 33
+      },
+      {
+        "id" : "withdraw_btn",
+        "type" : "JButton",
+        "title" : "Withdraw",
+        "x" : 448,
+        "y" : 20,
+        "width":106,
+        "height": 33
+      },
+      {
+        "id" : "add_interest_btn",
+        "type" : "JButton",
+        "title" : "Add interest",
+        "x" : 448,
+        "y" : 164,
+        "width":96,
+        "height": 33
+      },
+      {
+        "id" : "exit",
+        "type" : "JButton",
+        "title" : "Exit",
+        "x" : 448,
+        "y" : 248,
+        "width":96,
+        "height": 33
+      },
+      {
+        "id" : "scroll_panel",
+        "type" : "JScrollPanel",
+        "title" : "Exit",
+        "x" : 12,
+        "y" : 92,
+        "width":444,
+        "height": 160,
+        "children": [
+          {
+            "id" : "acc_table",
+            "type" : "JTable",
+            "title" : "",
+            "x" : 448,
+            "y" : 248,
+            "width":96,
+            "height": 33,
+            "columns" : ["AccountNr", "Name", "City", "P/C", "Ch/S", "Amount"]
+          }
+        ]
+      }
+    ]
+ }
